Rename 'users' to 'user' in some instances
