+++ exercises/03-redux/src/modules/root/components/root.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { UsersList } from 'modules/users/components/users-list'
+import { UserList } from 'modules/users/components/user-list'
 import { User, AddUserFunc } from 'modules/users/user-types'
 
 import { Header } from './header'
@@ -14,6 +14,6 @@ interface RootProps {
 export const Root: React.FC<RootProps> = ({ title, users, addUser }) => (
   <div>
     <Header title={title} />
-    <UsersList users={users} addUser={addUser} />
+    <UserList users={users} addUser={addUser} />
   </div>
 )
