+++ exercises/03-redux/src/modules/users/users-reducer.ts
@@ -1,4 +1,4 @@
-import * as UserActions from 'modules/users/users-actions'
+import * as UserActions from 'modules/users/user-actions'
 import { User } from './user-types'
 
 import { Reducer } from 'redux'
